requirements.txt
pandas==1.2.4
Flask==2.0.0
Flask_PyMongo==2.3.0
numpy==1.20.3
pymongo==3.11.4
Flask_Mail==0.9.1
python_bcrypt==0.3.2
gunicorn